+++ MDictUtils/MDictWriter.cs
@@ -4,7 +4,7 @@ using MDictUtils.BuildModels;
 
 namespace MDictUtils;
 
-public sealed record MDictEntry(string Key, long Pos, string Path, long Size)
+public sealed record MDictEntry(string Key, long Pos, string Path, int Size)
 {
     public override string ToString()
         => $"Key=\"{Key}\", Pos={Pos}, Size={Size}";
